+++ components/Projects/ProjectPreview.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { motion } from 'framer-motion';
 import gsap from 'gsap';
 import Image from 'next/image';
@@ -32,39 +34,41 @@ export default function ProjectPreview({ modal, projects }: ProjectModalProps) {
   const cursorLabel = useRef<HTMLDivElement>(null);
 
   useEffect(() => {
-    // Move Container
-    const xMoveContainer = gsap.quickTo(modalContainer.current, 'left', {
-      duration: 0.8,
-      ease: 'power3',
-    });
-    const yMoveContainer = gsap.quickTo(modalContainer.current, 'top', {
-      duration: 0.8,
-      ease: 'power3',
-    });
+    if (window !== undefined) {
+      // Move Container
+      const xMoveContainer = gsap.quickTo(modalContainer.current, 'left', {
+        duration: 0.8,
+        ease: 'power3',
+      });
+      const yMoveContainer = gsap.quickTo(modalContainer.current, 'top', {
+        duration: 0.8,
+        ease: 'power3',
+      });
 
-    // Move cursor
-    const xMoveCursor = gsap.quickTo(cursor.current, 'left', { duration: 0.5, ease: 'power3' });
-    const yMoveCursor = gsap.quickTo(cursor.current, 'top', { duration: 0.5, ease: 'power3' });
+      // Move cursor
+      const xMoveCursor = gsap.quickTo(cursor.current, 'left', { duration: 0.5, ease: 'power3' });
+      const yMoveCursor = gsap.quickTo(cursor.current, 'top', { duration: 0.5, ease: 'power3' });
 
-    // Move cursor label
-    const xMoveCursorLabel = gsap.quickTo(cursorLabel.current, 'left', {
-      duration: 0.45,
-      ease: 'power3',
-    });
-    const yMoveCursorLabel = gsap.quickTo(cursorLabel.current, 'top', {
-      duration: 0.45,
-      ease: 'power3',
-    });
+      // Move cursor label
+      const xMoveCursorLabel = gsap.quickTo(cursorLabel.current, 'left', {
+        duration: 0.45,
+        ease: 'power3',
+      });
+      const yMoveCursorLabel = gsap.quickTo(cursorLabel.current, 'top', {
+        duration: 0.45,
+        ease: 'power3',
+      });
 
-    window.addEventListener('mousemove', (e) => {
-      const { pageX, pageY } = e;
-      xMoveContainer(pageX);
-      yMoveContainer(pageY);
-      xMoveCursor(pageX);
-      yMoveCursor(pageY);
-      xMoveCursorLabel(pageX);
-      yMoveCursorLabel(pageY);
-    });
+      window.addEventListener('mousemove', (e) => {
+        const { pageX, pageY } = e;
+        xMoveContainer(pageX);
+        yMoveContainer(pageY);
+        xMoveCursor(pageX);
+        yMoveCursor(pageY);
+        xMoveCursorLabel(pageX);
+        yMoveCursorLabel(pageY);
+      });
+    }
   }, []);
 
   return (
