+++ components/Spotify/NowPlaying.tsx
@@ -35,7 +35,7 @@ async function fetchNowPlaying(): Promise<NowPlayingSong | null> {
   return null;
 }
 
-export const revalidate = 30;
+// export const revalidate = 30;
 
 export default async function NowPlaying() {
   const nowPlaying = await fetchNowPlaying();
